#177 moved checkemail file into helpers dir
