+++ src/controllers/helpers/check_email.js
