+++ src/controllers/send.js
@@ -1,5 +1,5 @@
 
-const checkEmail = require('./check_email');
+const checkEmail = require('./helpers/check_email');
 const sendemail = require('sendemail');
 const env = require('env2')('config.env');
